Rename file to match coding conventions.
