+++ src/qcmatlab/__init__.py
@@ -2,7 +2,7 @@
 # from qctoolkit.qcmatlab import engine
 # engine.workspace['max_variable'] = 42
 # engine.workspace['foo'] = 'bar'
-from .Manager import Manager
+from .manager import Manager
 
 engine = Manager.getEngine()
 connect = Manager.connectTo
